Split up CellMeasurer example files for better organization and readability
