+++ source/CellMeasurer/CellMeasurer.DynamicHeightList.example.js
@@ -1,0 +1,83 @@
+/** @flow */
+import Immutable from 'immutable'
+import React, { PropTypes, PureComponent } from 'react'
+import CellMeasurer from './CellMeasurer'
+import CellMeasurerCache from './CellMeasurerCache'
+import List from '../List'
+import styles from './CellMeasurer.example.css'
+
+export default class DynamicHeightList extends PureComponent {
+  static propTypes = {
+    getClassName: PropTypes.func.isRequired,
+    list: PropTypes.instanceOf(Immutable.List).isRequired,
+    width: PropTypes.number.isRequired
+  }
+
+  constructor (props, context) {
+    super(props, context)
+
+    this._cache = new CellMeasurerCache({
+      fixedWidth: true,
+      minHeight: 50
+    })
+
+    this._rowRenderer = this._rowRenderer.bind(this)
+  }
+
+  render () {
+    const { width } = this.props
+
+    return (
+      <List
+        className={styles.BodyGrid}
+        deferredMeasurementCache={this._cache}
+        height={400}
+        overscanRowCount={0}
+        rowCount={1000}
+        rowHeight={this._cache.rowHeight}
+        rowRenderer={this._rowRenderer}
+        width={width}
+      />
+    )
+  }
+
+  _rowRenderer ({ index, isScrolling, key, parent, style }) {
+    const { getClassName, list } = this.props
+
+    const datum = list.get(index % list.size)
+    const classNames = getClassName({ columnIndex: 0, rowIndex: index })
+
+    const imageWidth = 300
+    const imageHeight = datum.size
+
+    const source = `http://lorempixel.com/${imageWidth}/${imageHeight}/`
+
+    return (
+      <CellMeasurer
+        cache={this._cache}
+        columnIndex={0}
+        key={key}
+        parent={parent}
+        rowIndex={index}
+      >
+        {({ measure }) => (
+          <div
+            className={classNames}
+            style={{
+              ...style,
+              height: 'auto'
+            }}
+          >
+            <img
+              onLoad={measure}
+              src={source}
+              style={{
+                width: imageWidth
+              }}
+            />
+          </div>
+        )}
+      </CellMeasurer>
+    )
+  }
+}
